requirements.txt
# Core dependencies for AI trading system
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
torch==2.0.1
xgboost==1.7.6
lightgbm==4.0.0

# Data and API handling
yfinance==0.2.18
alpha-vantage==2.3.1
ccxt==4.0.30
websocket-client==1.6.1
requests==2.31.0
beautifulsoup4==4.12.2

# Database and caching
psycopg2-binary==2.9.7
sqlalchemy==2.0.19
redis==4.6.0
asyncpg==0.28.0

# Trading and technical analysis
talib==0.4.26
ta==0.10.2
backtrader==1.9.78.123
zipline-reloaded==3.0.3

# Async and messaging
asyncio==3.4.3
aiohttp==3.8.5
celery==5.3.1
rabbitmq==0.2.0

# Monitoring and logging
prometheus-client==0.17.1
grafana-api==1.0.3
structlog==23.1.0
sentry-sdk==1.29.2

# Configuration and utilities
pydantic==2.1.1
python-dotenv==1.0.0
click==8.1.6
rich==13.4.2
typer==0.9.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
mock==5.1.0

# Development
black==23.7.0
flake8==6.0.0
mypy==1.4.1
pre-commit==3.3.3

# Jupyter for analysis
jupyter==1.0.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Security and encryption
cryptography==41.0.3
bcrypt==4.0.1
pyjwt==2.8.0

# Time series and financial data
arch==6.2.0
pyfolio==0.9.2
empyrical==0.5.5
quantlib==1.31